feat: add new timelines and period
