+++ package.json
@@ -22,7 +22,7 @@
     "db:stop": "supabase stop",
     "db:sync": "supabase db pull",
     "db:test": "supabase test db",
-    "db:seed:electricity": "node --experimental-strip-types ./scripts/seed-electricity.ts",
+    "db:seed:electricity": "node --experimental-strip-types ./scripts/seed-electricity.mts",
     "db:gen:migration": "supabase migration new",
     "db:gen:types": "supabase gen types typescript --local > ./packages/services/src/supabase/types.ts",
     "prepare": "husky"
